requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
gevent==21.8.0
greenlet==1.1.1
grequests==0.6.0
idna==3.2
pycparser==2.20
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
requests==2.26.0
urllib3==1.26.6
zope.event==4.5.0
zope.interface==5.4.0

PyYAML~=5.4.1
twilio~=6.63.2